(feat) tsconfig feature for svelte-check (#1014)
Makes it possible to pass a path to a tsconfig or jsconfig file. The path can be relative to the workspace path or absolute. Doing this means that only files matched by the files/include/exclude pattern of the config file are diagnosed. It also means that errors from TypeScript and JavaScript files are reported.
#390

+++ packages/language-server/src/plugins/typescript/LSAndTSDocResolver.ts
@@ -4,20 +4,27 @@ import { LSConfigManager } from '../../ls-config';
 import { debounceSameArg, pathToUrl } from '../../utils';
 import { DocumentSnapshot, SvelteDocumentSnapshot } from './DocumentSnapshot';
 import {
-    getLanguageServiceForDocument,
-    getLanguageServiceForPath,
     getService,
+    getServiceForTsconfig,
+    hasServiceForFile,
     LanguageServiceContainer,
     LanguageServiceDocumentContext
 } from './service';
 import { SnapshotManager } from './SnapshotManager';
 
 export class LSAndTSDocResolver {
+    /**
+     *
+     * @param docManager
+     * @param workspaceUris
+     * @param configManager
+     * @param tsconfigPath This should only be set via svelte-check. Makes sure all documents are resolved to that tsconfig. Has to be absolute.
+     */
     constructor(
         private readonly docManager: DocumentManager,
         private readonly workspaceUris: string[],
         private readonly configManager: LSConfigManager,
-        private readonly transformOnTemplateError = true
+        private readonly tsconfigPath?: string
     ) {
         const handleDocumentChange = (document: Document) => {
             // This refreshes the document in the ts language service
@@ -55,12 +62,12 @@ export class LSAndTSDocResolver {
     private get lsDocumentContext(): LanguageServiceDocumentContext {
         return {
             createDocument: this.createDocument,
-            transformOnTemplateError: this.transformOnTemplateError
+            transformOnTemplateError: !this.tsconfigPath
         };
     }
 
     async getLSForPath(path: string) {
-        return getLanguageServiceForPath(path, this.workspaceUris, this.lsDocumentContext);
+        return (await this.getTSService(path)).getService();
     }
 
     async getLSAndTSDoc(document: Document): Promise<{
@@ -68,11 +75,7 @@ export class LSAndTSDocResolver {
         lang: ts.LanguageService;
         userPreferences: ts.UserPreferences;
     }> {
-        const lang = await getLanguageServiceForDocument(
-            document,
-            this.workspaceUris,
-            this.lsDocumentContext
-        );
+        const lang = await this.getLSForPath(document.getFilePath() || '');
         const tsDoc = await this.getSnapshot(document);
         const userPreferences = this.getUserPreferences(tsDoc.scriptKind);
 
@@ -93,6 +96,11 @@ export class LSAndTSDocResolver {
     }
 
     async deleteSnapshot(filePath: string) {
+        if (!hasServiceForFile(filePath, this.workspaceUris)) {
+            // Don't initialize a service for a file that should be deleted
+            return;
+        }
+
         (await this.getTSService(filePath)).deleteSnapshot(filePath);
         this.docManager.releaseDocument(pathToUrl(filePath));
     }
@@ -101,7 +109,13 @@ export class LSAndTSDocResolver {
         return (await this.getTSService(filePath)).snapshotManager;
     }
 
-    private getTSService(filePath: string): Promise<LanguageServiceContainer> {
+    async getTSService(filePath?: string): Promise<LanguageServiceContainer> {
+        if (this.tsconfigPath) {
+            return getServiceForTsconfig(this.tsconfigPath, this.lsDocumentContext);
+        }
+        if (!filePath) {
+            throw new Error('Cannot call getTSService without filePath and without tsconfigPath');
+        }
         return getService(filePath, this.workspaceUris, this.lsDocumentContext);
     }
